requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.1.0
django-cron==0.6.0
django-crontab==0.7.1
djangorestframework==3.14.0
whitenoise==6.4.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
